Extracted out loadCsv to Core.
