+++ lib/Core/loadCsv.ts
@@ -1,11 +1,17 @@
 import Papa from "papaparse";
-import loadText from "../../Core/loadText";
-import makeRealPromise from "../../Core/makeRealPromise";
+import loadText from "./loadText";
+import makeRealPromise from "./makeRealPromise";
 
+/**
+ * Fetches CSV from `url` and uses papaparse to parse it.
+ *
+ */
 export default function loadCsv(
   url: string,
   options?: Papa.ParseConfig
 ): Promise<any[]> {
+  // We use loadText instead of papaparse to fetch the data to follow the
+  // common practice of using Cesium Resource to do network requests.
   return makeRealPromise<string>(loadText(url)).then(
     (text: string) =>
       new Promise((resolve, reject) =>
